requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.20.0
azure-storage-blob==12.25.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
fastapi==0.115.11
gitdb==4.0.12
GitPython==3.1.44
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
narwhals==1.29.0
numpy==2.2.3
openai==1.65.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyJWT==2.10.1
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.23.1
schedule==1.2.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.0
streamlit==1.43.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn
watchdog==6.0.0
openpyxl==3.1.2
lxml==4.9.5
pyodbc